requirements.txt
# SHIOL+ Development Requirements
# Full dependencies for local development and testing

# Core Web Framework (pinned for reproducibility during dev)
fastapi==0.120.0
starlette==0.48.0
uvicorn[standard]==0.24.0
python-multipart==0.0.20

# Essential Utilities
requests==2.32.4
loguru==0.7.3
pytz==2023.3
jinja2==3.1.6
python-dotenv
beautifulsoup4==4.12.3

# Scheduling & Database
apscheduler==3.10.4
sqlalchemy==2.0.23

# Security & Authentication
bcrypt==4.1.2
psutil==5.9.8
PyJWT==2.8.0
email-validator==2.1.0

# Payment Processing
stripe==8.5.0

# ML Pipeline Dependencies
pandas==2.3.3
numpy==1.26.4
scikit-learn==1.7.0
xgboost==3.0.2
joblib==1.5.1
statsmodels==0.14.5
scipy==1.16.2

# AI Integration
google-generativeai==0.8.5
google-analytics-data==0.18.0

# Image Processing (Development)
Pillow==12.0.0
easyocr
google-cloud-vision
scikit-image

# Visualization (Development)
matplotlib==3.8.2
plotly==5.17.0